#27 - Fixed location problems
